\include is a dangerous command apparently. Switch to \input in all
Apparently \include is dangerous to some version of TeXLive.
This can be solved by only using the \input command because it ONLY
reads from the source file.

The -output-directory flag actually cd's us into the output directory,
meaning the \included files are in parent directories, which LaTeX/TeX
considers dangerous.