Phys 123: Basic Kinematics vector symbols' size sorted out
Some of the vector symbols were smaller than I would like, so I went
through and got the symbols to a size I do like.
However, the \left and \right commands do not work on the <> brackets
for vectors.
You can get around this with the \big, \Big, \bigg, and \Bigg commands.