Toyed with idea of putting examples in separate subdirs
There are 2 ways about this.
1. Using the \input{}s in the master document, like I am, you need to
provide the whole relative filepath to the example, from the main
document file.
2. Using TeXLive's pdflatex command cd's into the output directory,
causing some issues with \include.

There are a few potential solutions.
1. Just use the whole relative directory in each example.
2. Stop using the TeX_Output folder for storing PDFs, which will let
TeXLive's pdflatex use the \include command correctly.
3. Switch from TeXLive to mikTeX.

This will be very helpful for documents where there are a lot of examples.