EITF75: Started skeleton for 20190903 Lecture 2's typed notes
Lecture 2 was all about convolutions and how we can use and analyze
them.
This will be completed 20190904, because I ran out of time 20190903.
Also, I will be reading the section in the textbook about this topic,
so that might help clarify.
I also intend to watch some videos to make things even clearer.
This section will be finalized and finished when I complete Assignment
2.