Reference Sheet Preamble: Added file-local variables so emacs works
The TeX-master: shared variable means that the file will ask for the
master file, but it won't write it to the file.
This may mean that when I reopen Emacs after ending the task, I have
to redeclare the master for the preamble so I get the correct AucTeX
syntax highlighing, RefTeX completion, among everything else.