EDAN65: Added Semantic analysis, because we started abstract syntax
I need to fill in this section, but we started the abstract syntax
stuff, so we are moving through the compiler stuff to the semantic
analysis.
This means we start doing things like attaching types to parse nodes
and remove tokens that we don't care about once we created our tree.
For instance, if we have an ADD node, all we care about are the
operands, not the + operators, because we know it is an ADD node
already.
Lab 2 for EDAN65 was about parsing and this.