EDAN65: Separated AST and Semantic analysis
Technically, these 2 things are separate things. You create the AST
before you start your semantic analysis so the analysis is easier to
do.
The output parse tree from the parser is just too combersome to use in
the later stages of the compiler, so we make the AST and run with that instead.