EDAN65: Finished Aspect-Oriented Programming section
This section is pretty small because we didn't really discuss it too
much. We also did not do much with it in our lab assignmentss.
Assignment 3 used Visitors, and Assignment 4 threw us into Reference
Attribute Grammars. Thus, this section is somewhat sparse. It could be
filled in by suggestions from people, but I won't.