EITF75: Started equation sheet to print for the exam
The equation sheet will only have equations and properties that I
think I will need to quickly lookup any information I may need for the
exam.

This exam also allows any and all materials to be brought in from the
outside world, so long as it does not connect to the internet.