EDAP05: Made section for programming language history and made skeleton
That section is pretty massive in the textbook, and it'll take me
quite some time to read through and type it out.
Also, because it is not super important for this course, it will exist
as an appendix entry, but won't be in the appendix directory.
This is because this is fairly course-specific. It could be moved to
the appendix section later though.