EDAP05: Wrote "Are subclasses subtypes?"
Essentially, if I have a subclass, can I substitute it in for any
occurrence of the parent type/class? To implement this efficiently, we
must limit how we can extend an inherited class.
This is required to ensure there is type compatibility between the
parent superclass and the child subclass.