EDIN01: Included Berlekamp-Massey Algorithm
There is a blank spot for a Berlekamp-Massey Algorithm flowchart right
now. That is a job for tomorrow.

I also removed a double-definition for Block Ciphers. The definition
in the block ciphers section is much better than the one that I just
removed.
Plus, the overall size of the PDF went down too.