EITP20 ?s: Exchanged the article class for the exam class.
This makes it easier to typeset these as questions that I can answer
on the computer, then generate to get an answer sheet, while still
studying at the same time. I hope that this is slightly more
convenient and focusing than just going through the questions and
checking that I under stand them.