EDAP05: We can model the side effectful nature of UPDATING variables
Because UPDATING a variable is a state change in the variable's state,
we need a way to model it. It is not terribly useful from a functional
perspective, but explicit rules need to be made for the imperative
languages for how this update works.

The code example is to give an example of how a subexpression can
change the evaluation of a variable's update.