EDAF35: Wrote about syscall handler
The syscall handler is just a special kind of hardware exception
handler.
To get to the exception handler, the process must throw a
trap/exception up to the kernel, and the kernel must handle it. By
"handling" it, it actually calls the syscall handler, which then
carries out the syscall.