EDAF35: Reverse Frame/Page Lookup system for Page-Buffering
In this system, each frame also keeps a local history of which page
last inhabited it.
This way, if the page needs to be swapped out, then used before the
frame is overwritten, we can use the version that is still in the
frame, rather than having to swap the page back into memory.