EDAF35: What about some applications, like Databases?
Databases are very special constructs that know exactly the kind of
memory they need and use and how they interact with the storage
system.
Thus, using the OS's resources might actually hurt the performance of
this application, rather than help it, because the OS's algorithms are
meant to be used for general purposes; which the DB's algorithms are
specialized to itself.