EDAF35: Contiguous Allocation also has a problem with file size
Since the size of a file is unknown, there is no way to guarantee the
file will be allocated the correct amount of space that it will need
over its lifetime.
However, if we do know its destined size, but it takes a while to get
there, the allocated but unused space may cause us to waste space.