EDAF35: Another disadvantage is the required size for pointers
Thus, we introduce the Clustering solution, which groups a couple of
sectors in a track together, then makes a pointer between all of the
members of this cluster.
This reduces the number of pointers needed, overall, thus reducing the
amount of space that is required to store the pointers between the
blocks.
However, we also just gave up the ability to directly address the
blocks.