EDAF35: Indexed File Allocation still suffers from wasted space
This time, it could be the index that is the wasted space.
If a very small file, i.e. one that requires only a handful of data
blocks, means that the majority of the index block will just point to
NULL.

This also means that if we were more clever about how we went around
dealing with the index blocks, we might be able to get more storage
capacity out of the disk and its file system.