EDAF35: Equations to find maximum supported file size of UNIX inode
Flavius flat-out said that there would be a question about this on the
final exam. I asked him how to do it, and these were the equation
answers he hinted at.

To be fair, these are not all that difficult to guess and construct
yourself, since its just a bit of multiplication and addition.