EDAF35: Pointer sizes for blocks may represent an efficiency problem
Since 2^32 is not large enough to address all the storage we can have
today, especially not for single files.
Today, many files are regularly larger than 4GB, so we need to move up
to the next size.

But 2^64 can handle files that are SO large that it's ridiculous.
There are no drives that can hold a file large enough for 2^64 to be
too large a file. This comes at the cost of doubling the amount of
storage and memory we need to handle the pointers of the file system
structure.