EDAF35: Side benefit of Journaling FS is it responds faster to I/O
This is because sequential I/O to the Journal is being performed while
you are waiting for the "write" to happen, in addition, this is a
synchronous operation.
Then, in the background, the operating system will perform Journal
actions until it catches up with what is going on in the OS.

This is a lot faster than asynchronous demand I/O to the data directly.