EDAF35: "Solve" the problem of hash tables for directories
These are not real solutions, but help alleviate the problem.

Really, the key goal is to reduce the amount of searching necessary
for this thing to manage the directories and file entries of our
system. If we can get that down by splitting the entries across a hash
table, even with collisions, you are not likely to get a collision,
depending on the hashing algorithm used.