EDAF35: Introduce the idea of hardware Interrupts again
Remember, these are different than software interrupts, traps.
A syscall is a trap. The arguments are put on the stack, everything is
made "tidy", then the syscall/trap happens, which DOES cause a
hardware interrupt, but only to "make the processor listen".

Syscall traps that cause a hardware interrupt actually get a special
hardware interrupt number for them to be handled in they way that they want.