EDAF35: Motivate the use of a DMA controller
This is a slave controller built onto the CPU that handles the
bringing in of data from throughout the system.
The CPU dispatches the DMA to grab some information and returns to
some other execution.
The DMA then goes out and grabs it, puts it where the CPU specified,
and when it has gotten everything it was supposed to, interrupts the
CPU.
Then, the CPU can hand it off to where it needs to go.

This is helpful even though the DMA takes control of the memory bus,
because the overall gain in performance is worth it.