EDAF35: How the I/O subsystem handles spooling of data
Spooling is slowly collecting all the information for a logical unit into a
single file before sending the whole thing out at once.
The information is not necessarily generated all at once and not
necessarily all by one process.

By spooling, we create a complete image in memory of what we want to
end up with, then send it out.