EDAF35: How IO subsystem performance affects overall system performance
Overall, even a correctly implemented I/O subsystem can still be the
most limiting part of an OS, because that is just the nature of I/O.
Disks are incredibly slow relative to the processor and memory.
Network communication is even slower. Keyboard typing is even slower
than that.

In general, improvements to the I/O subsystem provide the largest
visible benefits to system response.

This is like upgrading an old computer from a regular HDD to SSD drive
for major system tasks.