Make sweeping change from \rightarrow to \to in approaching situations
There is a LaTeX command to describe the process of approaching a
value, whether it is in a limit, a summation, in-text, wherever.
I was not aware of this command for a long time, so I refactored many
of the \rightarrow's out to \to's.
There were some additional changes where \rightarrow's were changed to
\mapsto's, because that makes more sense than just an arrow.

Situations where there was a \Rightarrow, the => symbol, were left
unchanged, usually.