Reverted some \mapsto changes back to \to
Just like when programming in any strongly and strictly typed
programming language, like Haskell or Rust, you can infer the type of
a function from its definition.

In mathematics, without a definition, we cannot know the "type
signature" of the function. Thus, we have the syntax
  f: Input -> Output
which Haskell "borrowed" for its type signatures.

In commit dbcbdb924cb83302e578b3cb119ecd96e86ae1f6, I changed some of
the \rightarrow's there to \mapsto's, which was wrong.
I changed the appropriate function "type signatures" into \to's
because that is kind of what it sounds like when saying it out loud.

For example, f: \RealNumbers \to \RealNumbers is said
"The function f maps the set of all real numbers to the set of all
real numbers"